refactoring; enabled px, py, px option
