+++ code/train_script.py
@@ -10,11 +10,11 @@ from pathlib import Path
 from itertools import chain
 from torch.utils.data import random_split
 from torch_geometric.nn import EdgeConv, global_mean_pool, DataParallel
-from torch_geometric.data import Data, DataLoader, DataListLoader, Batch
+from torch_geometric.data import Data, DataLoader, DataListLoader
 
 import models.models as models
 import models.emd_models as emd_models
-from util.util import get_model
+from util.train_util import get_model, forward_loss
 from util.loss_util import LossFunction
 from datagen.graph_data_gae import GraphDataset
 from util.plot_util import loss_curves, plot_reco_difference, gen_in_out
@@ -22,37 +22,6 @@ from util.plot_util import loss_curves, plot_reco_difference, gen_in_out
 torch.manual_seed(0)
 device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
 multi_gpu = torch.cuda.device_count()>1
-
-# helper to perform correct loss
-def forward_and_loss(model, data, loss_ftn_obj):
-    
-    if not multi_gpu:
-        data = data.to(device)
-
-    if 'emd_loss' in loss_ftn_obj.name or loss_ftn_obj.name == 'chamfer_loss':
-        batch_output = model(data)
-        if multi_gpu:
-            data = Batch.from_data_list(data).to(device)
-        y = data.x
-        batch = data.batch
-        batch_loss = loss_ftn_obj.loss_ftn(batch_output, y, batch)
-
-    elif loss_ftn_obj.name == 'emd_in_forward':
-        _, batch_loss = model(data)
-        batch_loss = batch_loss.mean()
-
-    elif loss_ftn_obj.name == 'vae_loss':
-        batch_output, mu, log_var = model(data)
-        y = torch.cat([d.x for d in data]).to(device) if multi_gpu else data.x
-        y = y.contiguous()
-        batch_loss = loss_ftn_obj.loss_ftn(batch_output, y, mu, log_var)
-
-    else:
-        batch_output = model(data)
-        y = torch.cat([d.x for d in data]).to(device) if multi_gpu else data.x
-        y = y.contiguous()
-        batch_loss = loss_ftn_obj.loss_ftn(batch_output, y)
-    return batch_loss
 
 @torch.no_grad()
 def test(model, loader, total, batch_size, loss_ftn_obj):
@@ -62,7 +31,7 @@ def test(model, loader, total, batch_size, loss_ftn_obj):
     t = tqdm.tqdm(enumerate(loader),total=total/batch_size)
     for i,data in t:
 
-        batch_loss = forward_and_loss(model, data, loss_ftn_obj)
+        batch_loss = forward_loss(model, data, loss_ftn_obj, device, multi_gpu)
 
         batch_loss = batch_loss.item()
         sum_loss += batch_loss
@@ -79,7 +48,7 @@ def train(model, optimizer, loader, total, batch_size, loss_ftn_obj):
     for i,data in t:
         optimizer.zero_grad()
 
-        batch_loss = forward_and_loss(model, data, loss_ftn_obj)
+        batch_loss = forward_loss(model, data, loss_ftn_obj, device, multi_gpu)
         batch_loss.backward()
         optimizer.step()
